backend/requirements.txt
djangorestframework
markdown
django-filter
django-admin
mysqlclient
djangorestframework-simplejwt
python-dotenv
django-cors-headers